# lib/api-zod/src/generated/types/moveTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mission Control API
 * OpenAPI spec version: 0.1.0
 */
import type { MoveTaskBodyStatus } from "./moveTaskBodyStatus";

export interface MoveTaskBody {
  status: MoveTaskBodyStatus;
}